backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
anthropic==0.40.0
httpx==0.28.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
redis==5.2.1
orjson==3.10.12
geopy==2.4.1
numpy==2.2.0
scipy==1.14.1
aiosqlite==0.20.0
greenlet>=3.0.0
twilio>=9.0.0
